requirements.txt
asyncore-wsgi==0.0.9
mysql-connector-python==8.0.21
protobuf==3.13.0
pyodbc==4.0.30
six==1.15.0
sockets==1.0.0
